Fix wait calculation
Retrying only once is incorrect if we're set to wait for a given number
of ms, since there might be any arbitrary number of contenders.

Also, it was doing 'now - start' as the new wait time, which is incorrect.
Instead, calculate the proper end time, and continue retrying until that
time is reached, updating the wait time to be 'end - now' at each iteration.